merge Lecturecode/2025 09 05 (minus Service/Controller Adaptions) (#26)
* porting lecture code.

currently compiles. this seems to be the "security" "package" done, with changes here and there as required.

 this commits motto: "imports will continue until morale improves"

* more porting: the dto folder;

on itself does nothing ofc. therefore unsurprisingly it compiles (and still works on basic testing).

* more porting: config/swaggerconfig

seems to enable the "lock" icons in the swagger ui; compact

still compiles and seems to work

* more porting lecture code: it defines these 2 exceptions for later use.

can't tell if verbosity of declaration strictly necessary. but possibly not dangerous.

still compiles and seems to work